requirements.txt
openai==1.59.3
faiss-cpu==1.9.0.post1
streamlit==1.41.1
langgraph==0.2.60
langchain==0.3.14
langchain-openai==0.2.14
langchain-community==0.3.14
python-dotenv==1.0.1
PyMuPDF==1.25.1